Build Markov chain generators in the executor
Realized that these things block the main thread for quite a long time when they run. No idea if doing this will actually work since they're constructors though, guess we'll see!
